Extract common logic beteen inline and modal expression output
